minimal fix for cast from pointer to int of different size.

+++ xfree/xc/extras/Mesa/src/mesa/math/m_debug_xform.c
@@ -187,7 +187,7 @@ static int test_transform_function( transform_func func, int psize,
    mat->type = mtypes[mtype];
 
    m = mat->m;
-   ASSERT( ((GLuint)m & 15) == 0 );
+   ASSERT( ((GLuint)(long)m & 15) == 0 );
 
    init_matrix( m );
